Move fixtures into their own module; complete review suggestions
